backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
ezdxf==1.1.4
Pillow==10.1.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
